refactor: module 名を network から api モジュールに修正
