+++ core/api/proguard-rules.pro
