+++ core/api/src/main/AndroidManifest.xml
