+++ core/data/src/main/java/app/kaito_dogi/mybrary/core/data/repository/SearchBookRepositoryImpl.kt
@@ -1,11 +1,11 @@
 package app.kaito_dogi.mybrary.core.data.repository
 
+import app.kaito_dogi.mybrary.core.api.GoogleBooksApi
+import app.kaito_dogi.mybrary.core.api.response.model.ItemResponse
+import app.kaito_dogi.mybrary.core.api.response.model.convertor.toSearchResultBook
 import app.kaito_dogi.mybrary.core.domain.model.ExternalBookId
 import app.kaito_dogi.mybrary.core.domain.model.SearchResultBook
 import app.kaito_dogi.mybrary.core.domain.repository.SearchBookRepository
-import app.kaito_dogi.mybrary.core.network.api.GoogleBooksApi
-import app.kaito_dogi.mybrary.core.network.response.model.ItemResponse
-import app.kaito_dogi.mybrary.core.network.response.model.convertor.toSearchResultBook
 import javax.inject.Inject
 import javax.inject.Singleton
