--- core/network/src/main/java/app/kaito_dogi/mybrary/core/network/response/GetVolumeResponse.kt
@@ -1,5 +1,0 @@
-package app.kaito_dogi.mybrary.core.network.response
-
-import app.kaito_dogi.mybrary.core.network.response.model.ItemResponse
-
-typealias GetVolumeResponse = ItemResponse
