--- core/network/src/main/java/app/kaito_dogi/mybrary/core/network/retrofit/GoogleBooksRetrofit.kt
@@ -1,7 +1,0 @@
-package app.kaito_dogi.mybrary.core.network.retrofit
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-internal annotation class GoogleBooksRetrofit
