+++ settings.gradle.kts
@@ -16,12 +16,12 @@ dependencyResolutionManagement {
 
 rootProject.name = "Mybrary"
 include(":app")
+include(":core:api")
 include(":core:common")
 include(":core:data")
 include(":core:database:local")
 include(":core:designsystem")
 include(":core:domain")
-include(":core:network")
 include(":core:ui")
 include(":feature:mybookdetail")
 include(":feature:mybooklist")
